+++ package.json
@@ -2,7 +2,7 @@
   "name": "@jimber/pkid",
   "version": "1.0.2",
   "description": "Lightweight node client for pkid",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "standard; node test/test.js",
     "debug": "npm link @jimber/pkid && node --nolazy --inspect-brk=9229 test/test.js"
